back-end/requirements.txt
django==5.0.7
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-simplejwt[crypto]==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
psycopg==3.2.1
psycopg-c==3.2.1
psycopg-pool==3.2.2
python-decouple==3.8
django-cors-headers==4.4.0
drf-standardized-errors[openapi]==0.14.0
pillow==10.4.0